pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transcript-to-summary"
version = "0.1.0"
description = "Read a document and provide a summary using a local model."
requires-python = ">=3.9"
dependencies = [
  "sumy>=0.11.0",
  "nltk>=3.8",
  "python-docx>=1.0.0",
  "numpy>=1.26",
]

[project.scripts]
transcript-to-summary = "transcript_to_summary.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
